MAKE `TrackActionBuilderTest` & `TrackTrafficBuilderTest` test cases
